Integrate JaiaAbout window into panel structure
