+++ src/web/components/JaiaAbout/JaiaAbout.less
@@ -1,0 +1,12 @@
+.jaia-about {
+    position: fixed;
+
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+
+    height: 300px;
+    width: 300px;
+
+    background-color: white;
+}
