+++ src/web/components/JaiaAbout/JaiaAbout.tsx
@@ -1,0 +1,5 @@
+import "./JaiaAbout.less";
+
+export default function JaiaAbout() {
+    return <div className="jaia-about"></div>;
+}
